Refactor demo to androidApp
