+++ androidApp/.gitignore
