+++ androidApp/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
         android:theme="@style/Theme.ArtMaker"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.artmaker.androidApp.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.ArtMaker">
             <intent-filter>
