+++ androidApp/src/main/java/com/artmaker/androidApp/ui/theme/Type.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.artmaker.demo.ui.theme
+package com.artmaker.androidApp.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
